% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{industry_metadata}
\alias{industry_metadata}
\title{Industry Metadata}
\format{
A data frame with 9 rows and 3 columns:
\describe{
\item{code}{Character, industry code abbreviation}
\item{name_label}{Character, industry name}
\item{description}{Character, detailed industry description}
}
}
\usage{
industry_metadata
}
\description{
Contains a lookup table between industry codes and Chinese names,
along with detailed descriptions of each industry.
}
\examples{
# View industry metadata
data(industry_metadata)
print(industry_metadata)

}
\keyword{datasets}
